Support 0.5 case quantities across all pages
Items like chocolate sprinkles and M&Ms come in cases with one big bag,
so stores need to track half-case quantities. Changed all quantity inputs
to accept 0.5 increments and updated the database to use numeric columns.

+++ src/app/(authenticated)/items/actions.ts
@@ -20,8 +20,8 @@ export async function createItem(formData: FormData) {
     unit,
     base_unit: baseUnit,
     pcs_per_box: pcsPerBox,
-    reorder_point: parseInt(formData.get("reorder_point") as string) || 0,
-    target_stock: parseInt(formData.get("target_stock") as string) || 0,
+    reorder_point: parseFloat(formData.get("reorder_point") as string) || 0,
+    target_stock: parseFloat(formData.get("target_stock") as string) || 0,
   });
 
   if (error) return { error: error.message };
@@ -55,8 +55,8 @@ export async function updateItem(id: string, formData: FormData) {
       unit,
       base_unit: baseUnit,
       pcs_per_box: pcsPerBox,
-      reorder_point: parseInt(formData.get("reorder_point") as string) || 0,
-      target_stock: parseInt(formData.get("target_stock") as string) || 0,
+      reorder_point: parseFloat(formData.get("reorder_point") as string) || 0,
+      target_stock: parseFloat(formData.get("target_stock") as string) || 0,
     })
     .eq("id", id);
